rename otelcol to otlp
