+++ exporter/trace/otlp/transform_spans.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package otelcol
+package otlp
 
 import (
 	"google.golang.org/grpc/codes"
